+++ geopackage-ios/tiles/overlay/GPKGFeatureOverlayQuery.m
@@ -139,71 +139,6 @@
     return self.featureTiles.maxFeaturesPerTile != nil && tileFeaturesCount > [self.featureTiles.maxFeaturesPerTile intValue];
 }
 
--(GPKGBoundingBox *) buildClickBoundingBoxWithMapPoint: (GPKGMapPoint *) mapPoint andMapView: (MKMapView *) mapView{
-    return [self buildClickBoundingBoxWithLocationCoordinate:mapPoint.coordinate andMapView:mapView];
-}
-
--(GPKGBoundingBox *) buildClickBoundingBoxWithMKMapPoint: (MKMapPoint) mapPoint andMapView: (MKMapView *) mapView{
-    CLLocationCoordinate2D locationCoordinate = MKCoordinateForMapPoint(mapPoint);
-    return [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView];
-}
-
--(GPKGBoundingBox *) buildClickBoundingBoxWithPoint: (WKBPoint *) point andMapView: (MKMapView *) mapView{
-    CLLocationCoordinate2D locationCoordinate = CLLocationCoordinate2DMake([point.y doubleValue], [point.x doubleValue]);
-    return [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView];
-}
-
--(GPKGBoundingBox *) buildClickBoundingBoxWithLocationCoordinate: (CLLocationCoordinate2D) location andMapView: (MKMapView *) mapView{
-    CGPoint point = [mapView convertCoordinate:location toPointToView:mapView];
-    return [self buildClickBoundingBoxWithCGPoint:point andMapView:mapView];
-}
-
--(GPKGBoundingBox *) buildClickBoundingBoxWithCGPoint: (CGPoint) point andMapView: (MKMapView *) mapView{
-    
-    // Get the screen width and height a click occurs from a feature
-    int width = (int)roundf(mapView.frame.size.width * self.screenClickPercentage);
-    int height = (int)roundf(mapView.frame.size.height * self.screenClickPercentage);
-    
-    // Get the screen click locations in each width or height direction
-    CGPoint left = CGPointMake(point.x - width, point.y);
-    CGPoint up = CGPointMake(point.x, point.y - height);
-    CGPoint right = CGPointMake(point.x + width, point.y);
-    CGPoint down = CGPointMake(point.x, point.y + height);
-    
-    // Get the coordinates of the bounding box points
-    CLLocationCoordinate2D leftCoordinate = [mapView convertPoint:left toCoordinateFromView:mapView];
-    CLLocationCoordinate2D upCoordinate = [mapView convertPoint:up toCoordinateFromView:mapView];
-    CLLocationCoordinate2D rightCoordinate = [mapView convertPoint:right toCoordinateFromView:mapView];
-    CLLocationCoordinate2D downCoordinate = [mapView convertPoint:down toCoordinateFromView:mapView];
-    
-    GPKGBoundingBox * boundingBox = [[GPKGBoundingBox alloc] initWithMinLongitudeDouble:leftCoordinate.longitude
-                                                                  andMaxLongitudeDouble:rightCoordinate.longitude
-                                                                   andMinLatitudeDouble:downCoordinate.latitude
-                                                                   andMaxLatitudeDouble:upCoordinate.latitude];
-    
-    return boundingBox;
-}
-
--(GPKGBoundingBox *) buildClickBoundingBoxWithLocationCoordinate: (CLLocationCoordinate2D) location andMapBounds: (GPKGBoundingBox *) mapBounds{
-    
-    // Get the screen width and height a click occurs from a feature
-    struct GPKGBoundingBoxSize size = [mapBounds sizeInMeters];
-    double width = size.width * self.screenClickPercentage;
-    double height = size.height * self.screenClickPercentage;
-    
-    CLLocationCoordinate2D leftCoordinate = [GPKGTileBoundingBoxUtils locationWithBearing:270 andDistance:width fromLocation:location];
-    CLLocationCoordinate2D upCoordinate = [GPKGTileBoundingBoxUtils locationWithBearing:0 andDistance:height fromLocation:location];
-    CLLocationCoordinate2D rightCoordinate = [GPKGTileBoundingBoxUtils locationWithBearing:90 andDistance:width fromLocation:location];
-    CLLocationCoordinate2D downCoordinate = [GPKGTileBoundingBoxUtils locationWithBearing:180 andDistance:height fromLocation:location];
-    
-    GPKGBoundingBox * boundingBox = [[GPKGBoundingBox alloc] initWithMinLongitudeDouble:leftCoordinate.longitude
-                                                                  andMaxLongitudeDouble:rightCoordinate.longitude
-                                                                   andMinLatitudeDouble:downCoordinate.latitude
-                                                                   andMaxLatitudeDouble:upCoordinate.latitude];
-    
-    return boundingBox;
-}
-
 -(GPKGFeatureIndexResults *) queryFeaturesWithBoundingBox: (GPKGBoundingBox *) boundingBox{
     GPKGFeatureIndexResults * results = [self queryFeaturesWithBoundingBox:boundingBox withProjection:nil];
     return results;
@@ -455,7 +390,7 @@
     double zoom = [GPKGMapUtils currentZoomWithMapView:mapView];
     
     // Build a bounding box to represent the click location
-    GPKGBoundingBox * boundingBox = [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView];
+    GPKGBoundingBox * boundingBox = [GPKGMapUtils buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView andScreenPercentage:self.screenClickPercentage];
     
     NSString * message = [self buildMapClickMessageWithLocationCoordinate:locationCoordinate andZoom:zoom andClickBoundingBox:boundingBox andProjection:projection];
     
@@ -469,7 +404,7 @@
 -(NSString *) buildMapClickMessageWithLocationCoordinate: (CLLocationCoordinate2D) locationCoordinate andZoom: (double) zoom andMapBounds: (GPKGBoundingBox *) mapBounds andProjection: (GPKGProjection *) projection{
     
     // Build a bounding box to represent the click location
-    GPKGBoundingBox * boundingBox = [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapBounds:mapBounds];
+    GPKGBoundingBox * boundingBox = [GPKGMapUtils buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapBounds:mapBounds andScreenPercentage:self.screenClickPercentage];
     
     NSString * message = [self buildMapClickMessageWithLocationCoordinate:locationCoordinate andZoom:zoom andClickBoundingBox:boundingBox andProjection:projection];
     
@@ -527,7 +462,7 @@
     double zoom = [GPKGMapUtils currentZoomWithMapView:mapView];
     
     // Build a bounding box to represent the click location
-    GPKGBoundingBox * boundingBox = [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView];
+    GPKGBoundingBox * boundingBox = [GPKGMapUtils buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapView:mapView andScreenPercentage:self.screenClickPercentage];
     
     GPKGFeatureTableData * tableData = [self buildMapClickTableDataWithLocationCoordinate:locationCoordinate andZoom:zoom andClickBoundingBox:boundingBox andProjection:projection];
     
@@ -541,7 +476,7 @@
 -(GPKGFeatureTableData *) buildMapClickTableDataWithLocationCoordinate: (CLLocationCoordinate2D) locationCoordinate andZoom: (double) zoom andMapBounds: (GPKGBoundingBox *) mapBounds andProjection: (GPKGProjection *) projection{
     
     // Build a bounding box to represent the click location
-    GPKGBoundingBox * boundingBox = [self buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapBounds:mapBounds];
+    GPKGBoundingBox * boundingBox = [GPKGMapUtils buildClickBoundingBoxWithLocationCoordinate:locationCoordinate andMapBounds:mapBounds andScreenPercentage:self.screenClickPercentage];
     
     GPKGFeatureTableData * tableData = [self buildMapClickTableDataWithLocationCoordinate:locationCoordinate andZoom:zoom andClickBoundingBox:boundingBox andProjection:projection];